Added/fixed correct searchpath with $:.unshift in the examples
When you are testing xmpp4r from a checkout you want to use the checked out versions of the libs.
With $:.unshift '../../../../../lib' in every example the right libs get used, without it just the
global searchpath is used, which can lead to confusion when one script works and another doesn't.

Signed-off-by: Johannes Gilger <heipei@hackvalue.de>