Allow for running basic tarball only packaging tasks on systems without RubyGems.
WARNING : While you can run the rake tasks for simple packaging on systems
without RubyGems this is not recommended if you intend to also push changes
upstream to the Git repo.  The reason being that the rake tasks that autogenerate
the xmpp4r.gemspec will not be run on such systems and therefor any
add/remove/rename of files that are committed will not be reflected in the gemspec
and this can cause problems for other developers who wish to use the gem packaging
and who would now be out of sync.

If you intend only to run those package tasks locally (e.g. to build a .deb)
and not push file changes upstream from that repo that should not cause
any problems.