DRY up RDoc generation by sharing config between 'rake rdoc' and gem installation of ri/rdoc.
Now there is a single point of configuration in the Rakefile for rdoc generation when
created manually or as part of a gem install.

Commented out the --diagram, --fileboxes options for rdoc since they introduce a
hard binary dependency that is non-trivial to resolve on many systems and greatly slows
down the generation of rdocs.  This can be uncommented by devs who desire to see the DOT
graphs or for someone publishing the rdocs to the project website.

Commented out the --all option, and made it available for devs if desired.  In general
we should probably not generate the rdocs for private methods as we don't want to
encourage their use by non-devs.