Pubsub: fix namespace for purge_items_from
Received from lfcipriani on github.com:
--------------------
Hi Lucas,

I found a simple but annoying bug on pubsub support of xmpp4r, I think
that
isn't worthy to clone and send a push request to you. So, here it is:

On method purge_items_from of service_helper.rb

      def purge_items_from(node)
        iq = basic_pubsub_query(:set)
        purge = REXML::Element.new('purge')
        purge.attributes['node'] = node
        iq.pubsub.add(purge)
        @stream.send_with_id(iq)
      end

The call to method basic_pubsub_query(:set) should be
basic_pubsub_query(:set, true), because the namespace of this command is
pubsub#owner.

That's it. Thanks for this really useful gem.
--------------------