Add option to allow timeout to be specified for Jabber::Stream#send, use in Jabber::SASL::DigestMD5#auth on initial send.
Emperical evidence of a hang with Openfire 3.6.4 at that point; since the callback code is extremely simple, it points to not receiving a stanza and therefore not triggering the ThreadBlock#wakeup or ThreadBlock#raise, which hits a (long) mutex timeout.  Instead, set a 1s timeout on that send getting a response, try that up to 3 times, otherwise quit with a timeout exception.